+++ Demo/Sources/Demos/Demo Screens/SwipeActionsViewController.swift
@@ -16,7 +16,7 @@ final class SwipeActionsViewController: UIViewController  {
 
     private var allowDeleting: Bool = true
 
-    private var items = (0..<20).map { SwipeActionItem(isSaved: Bool.random(), identifier: $0) }
+    private var items = (0..<10).map { SwipeActionItem(isSaved: Bool.random(), identifier: $0) }
 
     override func loadView() {
         self.title = "Swipe Actions"
@@ -52,10 +52,29 @@ final class SwipeActionsViewController: UIViewController  {
                 }
             }
 
-            list += Section("items") { section in
+            list += Section("standardSwipeActionItems") { section in
+                section.header = DemoHeader(title: "Standard Style Swipable Items")
                 section += self.items.map { item in
                     Item(
-                        SwipeActionsDemoItem(item: item),
+                        SwipeActionsDemoItem(item: item, swipeActionsStyle: .default),
+                        swipeActions: self.makeSwipeActions(for: item)
+                    )
+                }
+            }
+
+            list += Section("customSwipeActionItems") { section in
+                section.header = DemoHeader(title: "Custom Style Swipable Items")
+                section += self.items.map { item in
+                    Item(
+                        SwipeActionsDemoItem(
+                            item: item,
+                            swipeActionsStyle:
+                                .init(
+                                    actionShape: .rectangle(cornerRadius: 8),
+                                    interActionSpacing: 8,
+                                    containerInsets: .init(top: 8, left: 8, bottom: 8, right: 8)
+                                )
+                        ),
                         swipeActions: self.makeSwipeActions(for: item)
                     )
                 }
@@ -128,6 +147,7 @@ final class SwipeActionsViewController: UIViewController  {
 
     struct SwipeActionsDemoItem: BlueprintItemContent, Equatable {
         var item: SwipeActionItem
+        var swipeActionsStyle: DefaultSwipeActionsView.Style
 
         var identifierValue: Int {
             self.item.identifier