+++ ListableUI/Sources/Internal/PresentationState/PresentationState.ItemState.swift
@@ -294,7 +294,7 @@ extension PresentationState
                         
             // Apply Swipe To Action Appearance
             if let actions = self.model.swipeActions {
-                cell.contentContainer.registerSwipeActionsIfNeeded(actions: actions, reason: reason)
+                cell.contentContainer.registerSwipeActionsIfNeeded(actions: actions, style: model.content.swipeActionsStyle, reason: reason)
             } else {
                 cell.contentContainer.deregisterSwipeIfNeeded()
             }