+++ ListableUI/Sources/Item/ItemContent.swift
@@ -353,7 +353,10 @@ public protocol ItemContent where Coordinator.ItemContentType == Self
     /// The view type to use to render swipe actions (delete, etc) for this content.
     /// A default implementation, which matches `UITableView`, is provided.
     associatedtype SwipeActionsView: ItemContentSwipeActionsView = DefaultSwipeActionsView
-    
+
+    /// Specify a swipe action style for this content.
+    var swipeActionsStyle: SwipeActionsView.Style { get }
+
     //
     // MARK: Creating & Providing Content Views
     //
@@ -437,7 +440,6 @@ public protocol ItemContent where Coordinator.ItemContentType == Self
     func makeCoordinator(actions : CoordinatorActions, info : CoordinatorInfo) -> Coordinator
 }
 
-
 /// The views owned by the item content, passed to the `apply(to:) method to theme and provide content.`
 public struct ItemContentViews<Content:ItemContent>
 {
@@ -481,6 +483,11 @@ public struct ApplyItemContentInfo
     public var environment : ListEnvironment
 }
 
+public extension ItemContent where SwipeActionsView.Style == DefaultSwipeActionsView.Style {
+    var swipeActionsStyle: SwipeActionsView.Style {
+        return .default
+    }
+}
 
 public extension ItemContent where Self:Equatable
 {
@@ -599,10 +606,12 @@ public extension ItemContent where BackgroundView == UIView
 /// If you do so, you're completely responsible for creating and laying out the actions,
 /// as well as updating the layout based on the swipe state.
 public protocol ItemContentSwipeActionsView: UIView {
+    /// Swipe action styles (e.g. `standard`, `grouped`, etc.) that are supported by the view.
+    associatedtype Style: Equatable
 
     var swipeActionsWidth: CGFloat { get }
 
-    init(didPerformAction: @escaping SwipeAction.CompletionHandler)
+    init(style: Style, didPerformAction: @escaping SwipeAction.CompletionHandler)
 
     func apply(actions: SwipeActionsConfiguration)